stan_glm.fit.R: don't include unused code in line coverage
@bgoodri We can probably go further than just not including this in
line coverage stats and get rid of these lines. I don’t think we need
the code for weights here since any weights specified will be
interpreted as number of trials and won’t be passed to Stan as weights.
Also, is there ever a reason to specify an offset in a bernoulli model?