stan_jm appears to working, but slow
I think the version of stan_jm in the forked rstanarm is pretty much
working.

One serious bug fix was related to the pad_reTrms, which had been
changed in rstanarm and didn't work with stan_jm. This was because the
'reTrms$Ztlist' object has a different structure depending on whether it
is returned by a call to lme4::glFormula, or a call to lme4::glmer
followed by lme4::getME(fittedmodel, "Ztlist")

The print and summary methods appear to be working for stanjm objects
too.

I haven't checked all the other stanjm methods.