Bug fix, for multivariate JM with etavalue assoc
Inside one of the (m in 1:M) loops in jm.stan, the indexing for the
etavalue assoc was M instead of m, which meant the etavalue for one
submodel was repeated M times!! Should now correctly loop over the
etavalues for each of the submodels.

Also using prior*_aux was throwing error, since wasn't being set to NULL
in matched call.