Addition of posterior_traj function to rstanarm
Added posterior_traj function.

Also made changes to posterior_predict so that it can handle predictions
for one submodel of a stanjm object, without compromising the handling
of other stanreg objects.

Also changed the majority of stanjm-methods, so that they have an
additional argument 'm'. If NULL, then the method generally returns the
result for *all* submodels, whereas if m = some value then the method
will return the result for one submodel. The m argument can of course be
included when calling stanreg-methods too, where it will just be
ignored.