Allow lower level clustering to only apply to some longitudinal markers and not others
Made changes to the Stan code that means that the clust_mat (which
summarises information in the lower level units within a patient, for
use in the association structure) can be used for a subset of the
longitudinal submodels.

This means that lower level units clustered within the patient might
exist for some markers, but not others.