stan_mvmer, stan_jm: add documentation for formula/formulaLong
Clarify in the documentation that nested grouping factors (g1/g2, g1:g2) are not allowed, nor is the double bar notation (||). This could change if I go back to parsing the formula using lme4, but at the moment the formula is parsed manually, so only the single bar notation with non-nested grouping factors is allowed.