Tidy up prior_summary for stan_jm/mvmer
Some aspects of the `lkj()` prior were not getting returned in the `prior.info` element of stan_jm/mvmer objects. They should be included now. The printing of the `lkj()` prior is now correct too -- before it was missing the adjusted scales and the df for the student-t component of the prior (ie. the student-t prior on the SDs of the random effects).

The `stan_mvmer` objects also needed a small fix. The named elements of their `prior.info` was priorLong, priorLong_intercept, etc. I needed to remove the "Long" from these names so that they printed correctly when `prior_summary()` was called on a `stan_mvmer` object.